fix Authorization Court

+++ app/Http/Controllers/CourtController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\CourtRequest;
-use App\Services\Impl\CourtService;
+use App\Services\ICourtService;
 use App\Models\Court;
 use App\Traits\ApiResponse;
 use Illuminate\Http\JsonResponse;
@@ -12,10 +12,11 @@ use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 class CourtController extends Controller
 {
     use ApiResponse;
+    use AuthorizesRequests;
 
-    private CourtService $courtService;
+    private ICourtService $courtService;
 
-    public function __construct(CourtService $courtService) {
+    public function __construct(ICourtService $courtService) {
         $this->courtService = $courtService;
     }
 
@@ -24,6 +25,7 @@ class CourtController extends Controller
     }
 
     public function store(CourtRequest $request): JsonResponse {
+        $this->authorize('create', Court::class);
         return $this->successResponse($this->courtService->add($request), "Saved Court");
     }
 
@@ -32,6 +34,7 @@ class CourtController extends Controller
     }
 
     public function update(string $id, CourtRequest $request): JsonResponse {
+        $this->authorize('update', Court::class);
         $data = $this->courtService->update($id, $request);
         if (!$data) {
             return $this->errorResponse("Updated Court Failed", 500);
@@ -40,6 +43,7 @@ class CourtController extends Controller
     }
 
     public function delete(string $id): JsonResponse {
+        $this->authorize('delete', Court::class);
         $data = $this->courtService->delete($id);
         if (!$data) {
             return $this->errorResponse("Deleted Court Failed", 500);
